+++ blocks/init/src/Blocks/components/tracking-head/tracking-head.php
@@ -1,9 +1,0 @@
-<?php
-
-/**
- * Tracking codes in Head
- *
- * @package %g_namespace%
- */
-
- // Adding Code here.