+++ scripts/components/re-orderable/re-orderable.js
@@ -4,14 +4,7 @@ import { SortableItem } from './sortable-item';
 import { Control } from '../base-control/base-control';
 import { restrictToVerticalAxis, restrictToHorizontalAxis, restrictToParentElement } from '@dnd-kit/modifiers';
 
-import {
-	DndContext,
-	rectIntersection,
-	KeyboardSensor,
-	PointerSensor,
-	useSensor,
-	useSensors,
-} from '@dnd-kit/core';
+import { DndContext, rectIntersection, KeyboardSensor, PointerSensor, useSensor, useSensors } from '@dnd-kit/core';
 
 import {
 	arrayMove,
@@ -84,7 +77,7 @@ export const ReOrderable = (props) => {
 		useSensor(PointerSensor),
 		useSensor(KeyboardSensor, {
 			coordinateGetter: sortableKeyboardCoordinates,
-		})
+		}),
 	);
 
 	const handleDragEnd = (event) => {
@@ -123,7 +116,7 @@ export const ReOrderable = (props) => {
 
 	// Check for duplicates and reassign IDs if needed.
 	const allIds = items?.map(({ id }) => id) ?? [];
-	const hasDuplicates = (input) => (new Set(input))?.size !== input?.length;
+	const hasDuplicates = (input) => new Set(input)?.size !== input?.length;
 
 	if (hasDuplicates(allIds) && items?.length > 0) {
 		const newItems = [...items].map((item, index) => ({ ...item, id: index + 1 }));
@@ -164,12 +157,13 @@ export const ReOrderable = (props) => {
 							icon={item?.props?.icon}
 							title={item?.props?.title ?? __('New item', 'eightshift-frontend-libs')}
 							subtitle={item?.props?.subtitle}
-							onRemove={item?.props?.onRemove ?? (
-								() => {
+							onRemove={
+								item?.props?.onRemove ??
+								(() => {
 									const newArray = [...items].filter((_, index) => index !== i);
 									setAttributes({ [attributeName]: newArray });
-								}
-							)}
+								})
+							}
 							isActive={items?.[i]?.id === active?.id}
 							isFirst={i === 0}
 							isLast={i === items?.length - 1}