+++ scripts/editor/editor.js
@@ -22,9 +22,7 @@ import { checkAttrResponsive, getAttrKey } from './attributes';
  * ```
  */
 export const lockPostEditing = (blockClientId, attributeKey) => {
-	dispatch('core/editor').lockPostSaving(
-		`undefined-lock-${blockClientId}-${attributeKey}`
-	);
+	dispatch('core/editor').lockPostSaving(`undefined-lock-${blockClientId}-${attributeKey}`);
 };
 
 /**
@@ -46,9 +44,7 @@ export const lockPostEditing = (blockClientId, attributeKey) => {
  * ```
  */
 export const unlockPostEditing = (blockClientId, attributeKey) => {
-	dispatch('core/editor').unlockPostSaving(
-		`undefined-lock-${blockClientId}-${attributeKey}`
-	);
+	dispatch('core/editor').unlockPostSaving(`undefined-lock-${blockClientId}-${attributeKey}`);
 };
 
 /**
@@ -98,11 +94,7 @@ export const lockIfUndefined = (blockClientId, attributeKey, value) => {
  * @returns React.Component
  */
 export const GutenbergBlock = (props) => {
-	const {
-		options: OptionsComponent,
-		toolbar: ToolbarComponent,
-		editor: EditorComponent,
-	} = props;
+	const { options: OptionsComponent, toolbar: ToolbarComponent, editor: EditorComponent } = props;
 
 	return (
 		<>
@@ -136,17 +128,12 @@ export const GutenbergBlock = (props) => {
  *
  * @returns Object
  */
-export const getResponsiveLegacyData = (
-	responsiveAttr,
-	attributes,
-	manifest,
-	setAttributes
-) => ({
+export const getResponsiveLegacyData = (responsiveAttr, attributes, manifest, setAttributes) => ({
 	attribute: Object.fromEntries(
 		Object.entries(responsiveAttr).map(([breakpoint, attrName]) => [
 			breakpoint,
 			getAttrKey(attrName, attributes, manifest),
-		])
+		]),
 	),
 	value: attributes,
 	onChange: (attributeName, value) => setAttributes({ [attributeName]: value }),