+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [15.0.0]
+
+### Changed
+
+- Webpack config now user mjs files and no longer has assets config key.
+- Es and style linters
+
+### Removed
+
+- Most of the SCSS mixing and files
+- All initial Blocks
+
 ## [14.0.1]
 
 ### Changed
@@ -1421,6 +1433,7 @@ Follow this migration script in order for you project to work correctly with the
 
 - Initial tagged release.
 
+[15.0.0]: https://github.com/infinum/eightshift-frontend-libs/compare/14.0.1...15.0.0
 [14.0.1]: https://github.com/infinum/eightshift-frontend-libs/compare/14.0.0...14.0.1
 [14.0.0]: https://github.com/infinum/eightshift-frontend-libs/compare/13.0.11...14.0.0
 [13.0.11]: https://github.com/infinum/eightshift-frontend-libs/compare/13.0.10...13.0.11